phyx now using double precision so collisions and high speed is sexy

+++ ENG/objects/kldr.cpp
@@ -11,7 +11,7 @@ bool CircleCollider::boolin(CircleCollider g){
 	return (glm::distance(pos,g.pos) <= (dim+g.dim));
 }
 
-bool CircleCollider::boolin(glm::vec2 other){
+bool CircleCollider::boolin(glm::dvec2 other){
 	return (
 		glm::distance(this->pos,other)
 		<= this->dim);
